@@ -1,7 +1,7 @@
 package com.tlam.miammiamk
 
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.widget.ImageView
 import android.widget.TextView
 import com.bumptech.glide.Glide
