@@ -6,10 +6,10 @@ import android.content.Intent
 import android.content.ServiceConnection
 import android.os.Bundle
 import android.os.IBinder
-import android.support.v7.app.AppCompatActivity
-import android.support.v7.widget.LinearLayoutManager
-import android.support.v7.widget.RecyclerView
-import android.support.v7.widget.Toolbar
+import androidx.appcompat.app.AppCompatActivity
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
+import androidx.appcompat.widget.Toolbar
 import android.util.Log
 import android.view.Menu
 import android.view.MenuItem
@@ -28,7 +28,7 @@ class MainActivity : AppCompatActivity() {
     var adapter: CuisineRecyclerViewAdapter? = null
 
     companion object {
-        val tag = "MainActivity"
+        const val tag = "MainActivity"
         var ctx: Context? = null
         var toolbar: Toolbar? = null
     }
